Move Currency and PaymentType enums to api module
